Uses unbound window functions directly
This is to fix errors that occur on sites that override the `window` global
with a Proxy'd version of their own.

Specifically, this happens on sites embedded on https://usepastel.com

See the files at https://github.com/romellem/temp-site/tree/master/embla-carousel
for more information on this.

If it is still up and running, the above repo is embedded at

    https://usepastel.com/link/2w7yy/

Otherwise, simply clone that repo, spin up a GitHub Pages site,
create a free account on Pastel, and create a new "canvas"
from your new site. Navigate to the `works.html` and `fails.html`
pages to view the errors.