Add a hint for flexbox and autoheight compatability
When following along with the Get Started guide, the example CSS doesn't
allow autoheight to work properly. The flex container will set all the
slides to be the same height making the autoheight script useless.
Adding some direction to the docs to amend the example CSS will help
avoid this issue.